The last row in JSON was not read
because JSON and CSV have different logic for header values the last JSON row was not read. Now it is done that when reading CSV tthe first time we read the columns rows is back to 0. This might need a further change in case our table doesn't have headers. This could happen with a checkbox that asks the user to check it if the table to be presented in CSV does not have headers.