Reading file dynamically from Open file
The table is cleared when opening a file from the File explorer. Known issue, if we reload the hard coded JSON and CSV files several time under the same program, they will append or be displayed at the bottom. Have to find a way to clear the observable after I am done with reading these files.